requirements.txt
flask
numpy
librosa
matplotlib
pillow
scikit-learn
sounddevice
soundfile
